chore(cli+persistence): implement requested changes
