+++ src/main/kotlin/hwr/oop/tnp/cli/TotallyNotPokemon.kt
@@ -1,175 +1,261 @@
 package hwr.oop.tnp.cli
 
 import hwr.oop.tnp.core.Attack
-import hwr.oop.tnp.core.GameUsage
 import hwr.oop.tnp.core.PrimitiveType
-import hwr.oop.tnp.persistency.GamePersistencePort
 import hwr.oop.tnp.persistency.FileSystemBasedJSONPersistence
+import hwr.oop.tnp.persistency.LoadBattlePort
 
 class TotallyNotPokemon(
-  private val args: List<String>,
-  private val adapter: GamePersistencePort = FileSystemBasedJSONPersistence(),
+        private val args: List<String>,
+        private val loadAdapter: LoadBattlePort = FileSystemBasedJSONPersistence(),
 ) {
-  private val game: GameUsage = Game()
+        private val COULD_NOT_PARSE_ERROR =
+                "Some of the provided arguments could not be parsed correctly"
+        private val COULD_NOT_PARSE_TO_INT_ERROR =
+                "Some of the provided arguments could not be parsed to an Int"
 
-  init {
-    parseArguments()
-  }
+        private val defaultHelp =
+                """.___________  _____  ___     _______
+("     _   ")(\"   \|"  \   |   __ "\
+ )__/  \\__/ |.\\   \    |  (. |__) :)
+    \\_ /    |: \.   \\  |  |:  ____/
+    |.  |    |.  \    \. |  (|  /
+    \:  |    |    \    \ | /|__/ \
+     \__|     \___|\____\)(_______)
 
-  fun parseArguments() {
-    if (args.isEmpty()) {
-      printHelp()
-      return
-    }
+Totally Not Pokémon Usage: ./tnp command [arguments]
+   Commands:
+      new_trainer     - Creates a new trainer
+      add_monster     - Adds a new monster to your roster
+      new_battle      - Starts a battle sequence
+      view_battle     - Views current game status
+      on              - Select a attack to perform
+      help            - Shows this help message"""
 
-    val command = args[0]
-    val arguments = args.slice(1..args.size - 1)
-    when (command) {
-      commands[0] -> prepareForCreateTrainer(arguments)
-      commands[1] -> parseForAddMonster(arguments)
-      commands[2] -> parseForNewBattle(arguments)
-      commands[3] -> parseForViewBattle(arguments)
-      commands[4] -> parseForPerformAttack(arguments)
-      commands[5] -> {
-        if (args.size > 1) {
-          printHelp(args[1])
-        } else {
-          printHelp()
+        private val newTrainerHelp =
+                """._____          _                   _   _      _
+|_   _| __ __ _(_)_ __   ___ _ __  | | | | ___| |_ __
+  | || '__/ _` | | '_ \ / _ \ '__| | |_| |/ _ \ | '_ \
+  | || | | (_| | | | | |  __/ |    |  _  |  __/ | |_) |
+  |_||_|  \__,_|_|_| |_|\___|_|    |_| |_|\___|_| .__/
+                                                |_|
+
+Usage: ./tnp new_trainer <TRAINERNAME> <BATTLE_ID>"""
+
+        private val addMonsterHelp =
+                """.   _       _     _   __  __                 _              _   _      _
+   / \   __| | __| | |  \/  | ___  _ __  ___| |_ ___ _ __  | | | | ___| |_ __
+  / _ \ / _` |/ _` | | |\/| |/ _ \| '_ \/ __| __/ _ \ '__| | |_| |/ _ \ | '_ \
+ / ___ \ (_| | (_| | | |  | | (_) | | | \__ \ ||  __/ |    |  _  |  __/ | |_) |
+/_/   \_\__,_|\__,_| |_|  |_|\___/|_| |_|___/\__\___|_|    |_| |_|\___|_| .__/
+                                                                        |_|
+
+Usage: ./tnp add_monster <MONSTERNAME> <HP_VALUE> <SPEED_VALUE> <TYPE> <ATTACK 1> [<ATTACK 2> <ATTACK 3> <ATTACK 4>] <TRAINER> <BATTLE_ID>"""
+
+        private val newBattleHelp =
+                """._   _                 ____        _   _   _        _   _      _
+| \ | | _____      __ | __ )  __ _| |_| |_| | ___  | | | | ___| |_ __
+|  \| |/ _ \ \ /\ / / |  _ \ / _` | __| __| |/ _ \ | |_| |/ _ \ | '_ \
+| |\  |  __/\ V  V /  | |_) | (_| | |_| |_| |  __/ |  _  |  __/ | |_) |
+|_| \_|\___| \_/\_/   |____/ \__,_|\__|\__|_|\___| |_| |_|\___|_| .__/
+                                                                |_|
+
+Usage: ./tnp new_battle"""
+
+        private val viewBattleHelp =
+                """__     ___                 ____        _   _   _        _   _      _
+\ \   / (_) _____      __ | __ )  __ _| |_| |_| | ___  | | | | ___| |_ __
+ \ \ / /| |/ _ \ \ /\ / / |  _ \ / _` | __| __| |/ _ \ | |_| |/ _ \ | '_ \
+  \ V / | |  __/\ V  V /  | |_) | (_| | |_| |_| |  __/ |  _  |  __/ | |_) |
+   \_/  |_|\___| \_/\_/   |____/ \__,_|\__|\__|_|\___| |_| |_|\___|_| .__/
+                                                                    |_|
+
+Usage: ./tnp view_battle <BATTLE_ID> | ALL
+
+Examples:
+1. View a specific battle with ID 123:
+   - `./tnp view_battle 123`
+
+2. View all battles:
+   - `./tnp view_battle ALL`"""
+
+        private val attackHelp =
+                """.   _   _   _             _      _   _      _
+   / \ | |_| |_ __ _  ___| | __ | | | | ___| |_ __
+  / _ \| __| __/ _` |/ __| |/ / | |_| |/ _ \ | '_ \
+ / ___ \ |_| || (_| | (__|   <  |  _  |  __/ | |_) |
+/_/   \_\__|\__\__,_|\___|_|\_\ |_| |_|\___|_| .__/
+                                             |_|
+
+Usage: ./tnp on <BATTLE_ID> <ATTACKNAME>"""
+
+        private val commands: List<String> =
+                listOf("new_trainer", "add_monster", "new_battle", "view_battle", "on", "help")
+
+        private val commandsHelpMap: Map<String, String> =
+                mapOf(
+                        commands[0] to newTrainerHelp,
+                        commands[1] to addMonsterHelp,
+                        commands[2] to newBattleHelp,
+                        commands[3] to viewBattleHelp,
+                        commands[4] to attackHelp,
+                        commands[5] to defaultHelp,
+                )
+
+        lateinit private var cliAdapter: BattleCliAdapter
+
+        init {
+                parseArguments()
         }
-      }
 
-      else -> println("'$command' is not a valid command. Use 'help' for usage.")
-    }
-  }
+        fun parseArguments() {
+                if (args.isEmpty()) {
+                        printHelp()
+                        return
+                }
 
-  private fun parseToInt(argument: String): Int {
-    return try {
-      argument.toInt()
-    } catch (e: NumberFormatException) {
-      throw Exception(
-        "Error: Failed to convert '$argument' to Int. Reason: ${e.message}"
-      )
-    }
-  }
+                val command = args[0]
+                val arguments = args.slice(1..args.size - 1)
+                when (command) {
+                        commands[0] -> prepareForCreateTrainer(arguments)
+                        commands[1] -> parseForAddMonster(arguments)
+                        commands[2] -> parseForNewBattle(arguments)
+                        commands[3] -> parseForViewBattle(arguments)
+                        commands[4] -> parseForPerformAttack(arguments)
+                        commands[5] -> {
+                                if (args.size > 1) {
+                                        printHelp(args[1])
+                                } else {
+                                        printHelp()
+                                }
+                        }
+                        else -> println("'$command' is not a valid command. Use 'help' for usage.")
+                }
+        }
 
-  fun parseToAttack(input: String): Attack {
-    return try {
-      Attack.valueOf(input.uppercase())
-    } catch (e: IllegalArgumentException) {
-      throw Exception(
-        "Error: Failed to convert '$input' to Attack. Reason: ${e.message}"
-      )
-    }
-  }
+        private fun parseToInt(argument: String): Int {
+                return try {
+                        argument.toInt()
+                } catch (e: NumberFormatException) {
+                        throw Exception(
+                                "Error: Failed to convert '$argument' to Int. Reason: ${e.message}"
+                        )
+                }
+        }
 
-  fun parseToType(input: String): PrimitiveType {
-    return try {
-      PrimitiveType.valueOf(input.uppercase())
-    } catch (e: IllegalArgumentException) {
-      throw Exception(
-        "Error: Failed to convert '$input' to Type. Reason: ${e.message}"
-      )
-    }
-  }
+        fun parseToAttack(input: String): Attack {
+                return try {
+                        Attack.valueOf(input.uppercase())
+                } catch (e: IllegalArgumentException) {
+                        throw Exception(
+                                "Error: Failed to convert '$input' to Attack. Reason: ${e.message}"
+                        )
+                }
+        }
 
-  private fun prepareForCreateTrainer(args: List<String>) {
-    if (args.size != 2) {
-      println(newTrainerHelp)
-      return
-    }
+        fun parseToType(input: String): PrimitiveType {
+                return try {
+                        PrimitiveType.valueOf(input.uppercase())
+                } catch (e: IllegalArgumentException) {
+                        throw Exception(
+                                "Error: Failed to convert '$input' to Type. Reason: ${e.message}"
+                        )
+                }
+        }
 
-    try {
-      val battle = adapter.loadBattle(args[1])
-      game.createTrainer(args[0], battle)
-      adapter.saveBattle(battle)
-    } catch (e: IllegalArgumentException) {
-      println(e.message)
-    }
-  }
+        private fun prepareForCreateTrainer(args: List<String>) {
+                if (args.size != 2) {
+                        println(newTrainerHelp)
+                        return
+                }
 
-  private fun parseForAddMonster(args: List<String>) {
-    if (args.isEmpty() || !(args.size >= 7 && args.size <= 10)) {
-      println(addMonsterHelp)
-      return
-    }
+                try {
+                        cliAdapter = BattleCliAdapter(loadAdapter.loadBattle(args[1]))
+                        cliAdapter.createTrainer(args[0])
+                } catch (e: IllegalArgumentException) {
+                        println(e.message)
+                }
+        }
 
-    val monsterName = args[0]
-    val trainerName = args[args.size - 2]
+        private fun parseForAddMonster(args: List<String>) {
+                if (args.isEmpty() || !(args.size >= 7 && args.size <= 10)) {
+                        println(addMonsterHelp)
+                        return
+                }
 
-    try {
-      val hp = parseToInt(args[1])
-      val speed = parseToInt(args[2])
-      val type = parseToType(args[3])
-      val attackList: MutableList<Attack> = mutableListOf()
-      for (attack in args.slice(4..args.size - 3).toList()) {
-        attackList.add(parseToAttack(attack))
-      }
-      val battle = adapter.loadBattle(args[args.size - 1])
+                val monsterName = args[0]
+                val trainerName = args[args.size - 2]
 
-      game.addMonster(
-        monsterName,
-        hp,
-        speed,
-        type,
-        attackList,
-        trainerName,
-        battle
-      )
-      adapter.saveBattle(battle)
-    } catch (e: Exception) {
-      println(COULD_NOT_PARSE_ERROR)
-      return
-    }
-  }
+                try {
+                        val hp = parseToInt(args[1])
+                        val speed = parseToInt(args[2])
+                        val type = parseToType(args[3])
+                        val attackList: MutableList<Attack> = mutableListOf()
+                        for (attack in args.slice(4..args.size - 3).toList()) {
+                                attackList.add(parseToAttack(attack))
+                        }
+                        val battle = loadAdapter.loadBattle(args[args.size - 1])
 
-  private fun parseForNewBattle(args: List<String>) {
-    if (!args.isEmpty()) {
-      println(newBattleHelp)
-      return
-    }
-    adapter.saveBattle(game.initiateBattle())
-  }
+                        cliAdapter.addMonster(
+                                monsterName,
+                                hp,
+                                speed,
+                                type,
+                                attackList,
+                                trainerName,
+                        )
+                } catch (e: Exception) {
+                        println(COULD_NOT_PARSE_ERROR)
+                        return
+                }
+        }
 
-  private fun parseForViewBattle(args: List<String>) {
-    if (args.isEmpty() || args.size != 1) {
-      println(viewBattleHelp)
-      return
-    }
+        private fun parseForNewBattle(args: List<String>) {
+                if (!args.isEmpty()) {
+                        println(newBattleHelp)
+                        return
+                }
+        }
 
-    if (args[0].trim().lowercase() == "all") {
-      game.showAllBattles(adapter.loadAllBattles())
-      return
-    }
+        private fun parseForViewBattle(args: List<String>) {
+                if (args.isEmpty() || args.size != 1) {
+                        println(viewBattleHelp)
+                        return
+                }
 
-    try {
-      val battle = adapter.loadBattle(args[0])
-      game.viewStatus(battle)
-    } catch (e: IllegalArgumentException) {
-      println(e.message)
-    }
-  }
+                if (args[0].trim().lowercase() == "all") {
+                        BattleCliAdapter.showAllBattles(loadAdapter.loadAllBattles())
+                        return
+                }
 
-  private fun parseForPerformAttack(args: List<String>) {
-    if (args.isEmpty() || args.size != 2) {
-      println(attackHelp)
-      return
-    }
+                try {
+                        val battle = loadAdapter.loadBattle(args[0])
+                        cliAdapter.viewStatus(battle)
+                } catch (e: IllegalArgumentException) {
+                        println(e.message)
+                }
+        }
 
-    try {
-      val battle = adapter.loadBattle(args[0])
-      game.performAttack(battle, parseToAttack(args[1]))
-      adapter.saveBattle(battle)
-    } catch (e: IllegalArgumentException) {
-      println(e.message)
-    }
-  }
+        private fun parseForPerformAttack(args: List<String>) {
+                if (args.isEmpty() || args.size != 2) {
+                        println(attackHelp)
+                        return
+                }
 
-  private fun printHelp(command: String = "") {
-    val helpMsg = commandsHelpMap.getOrDefault(command, defaultHelp)
-    println(helpMsg)
-  }
+                try {
+                        cliAdapter = BattleCliAdapter(loadAdapter.loadBattle(args[0]))
+                        cliAdapter.performAttack(parseToAttack(args[1]))
+                } catch (e: IllegalArgumentException) {
+                        println(e.message)
+                }
+        }
+
+        private fun printHelp(command: String = "") {
+                val helpMsg = commandsHelpMap.getOrDefault(command, defaultHelp)
+                println(helpMsg)
+        }
 }
 
 fun main(args: Array<String>) {
-  TotallyNotPokemon(args.toList())
+        TotallyNotPokemon(args.toList())
 }
