+++ src/main/kotlin/hwr/oop/tnp/core/Battle.kt
@@ -1,128 +1,128 @@
 package hwr.oop.tnp.core
 
-import kotlinx.serialization.Serializable
+import hwr.oop.tnp.persistency.FileSystemBasedJSONPersistence
+import hwr.oop.tnp.persistency.SaveBattlePort
 import java.util.*
+import kotlinx.serialization.Serializable
 
 @Serializable
 class Battle(
-  val battleId: String = UUID.randomUUID().toString(),
-) {
-  var trainerOne: Trainer = Trainer.EMPTY
-    private set
-  var trainerTwo: Trainer = Trainer.EMPTY
-    private set
-  var currentTrainer: Trainer = Trainer.EMPTY
-    private set
+        val battleId: String = UUID.randomUUID().toString(),
+        private val saveAdapter: SaveBattlePort = FileSystemBasedJSONPersistence(),
+) : BattleUsage {
+        var trainerOne: Trainer = Trainer.EMPTY
+                private set
+        var trainerTwo: Trainer = Trainer.EMPTY
+                private set
+        var currentTrainer: Trainer = Trainer.EMPTY
+                private set
 
-  var status: BattleStatus = BattleStatus.PREGAME
-    private set
+        override var status: BattleStatus = BattleStatus.PREGAME
+                private set
 
-  var currentRound: Int = 1
-    private set
+        var currentRound: Int = 1
+                private set
 
-  companion object {
-    private fun determineBeginningTrainer(
-      trainerOne: Trainer,
-      trainerTwo: Trainer,
-    ): Trainer {
-      val m1 = trainerOne.nextMonster()
-      val m2 = trainerTwo.nextMonster()
+        companion object {
+                private fun determineBeginningTrainer(
+                        trainerOne: Trainer,
+                        trainerTwo: Trainer,
+                ): Trainer {
+                        val m1 = trainerOne.nextMonster()
+                        val m2 = trainerTwo.nextMonster()
 
-      return if (m1.isFasterThan(m2)) trainerOne else trainerTwo
-    }
-  }
+                        return if (m1.isFasterThan(m2)) trainerOne else trainerTwo
+                }
+        }
 
-  fun getTrainerByName(name: String): Trainer {
-    val one = trainerOne
-    val two = trainerTwo
+        override fun getTrainerByName(name: String): Trainer {
+                val one = trainerOne
+                val two = trainerTwo
 
-    return when {
-      one.name.equals(name, ignoreCase = true) -> one
-      two.name.equals(name, ignoreCase = true) -> two
-      else -> throw IllegalArgumentException("Trainer '$name' not found.")
-    }
-  }
+                return when {
+                        one.name.equals(name, ignoreCase = true) -> one
+                        two.name.equals(name, ignoreCase = true) -> two
+                        else -> throw IllegalArgumentException("Trainer '$name' not found.")
+                }
+        }
 
-  fun addTrainerToBattle(trainer: Trainer) {
-    when {
-      trainerOne == Trainer.EMPTY -> trainerOne = trainer
-      trainerTwo == Trainer.EMPTY -> trainerTwo = trainer
-      else -> throw IllegalStateException("Both trainers are already set.")
-    }
-  }
+        override fun addTrainerToBattle(trainer: Trainer) {
+                when {
+                        trainerOne == Trainer.EMPTY -> trainerOne = trainer
+                        trainerTwo == Trainer.EMPTY -> trainerTwo = trainer
+                        else -> throw IllegalStateException("Both trainers are already set.")
+                }
+                saveAdapter.saveBattle(this)
+        }
 
-  fun startBattle() {
-    requireNoTrainerIsEmpty()
-    currentTrainer = determineBeginningTrainer(trainerOne, trainerTwo)
-    status = BattleStatus.STARTED
-  }
+        override fun startBattle() {
+                requireNoTrainerIsEmpty()
+                currentTrainer = determineBeginningTrainer(trainerOne, trainerTwo)
+                status = BattleStatus.STARTED
+                saveAdapter.saveBattle(this)
+        }
 
-  private fun endBattle() {
-    status = BattleStatus.FINISHED
-  }
+        private fun endBattle() {
+                status = BattleStatus.FINISHED
+        }
 
-  private fun advanceRound() {
-    if (determineWinner() != Trainer.EMPTY) {
-      endBattle()
-      return
-    }
-    currentRound++
-  }
+        private fun advanceAndSaveRound() {
+                if (determineWinner() != Trainer.EMPTY) {
+                        endBattle()
+                        return
+                }
+                currentRound++
+                saveAdapter.saveBattle(this)
+        }
 
-  private fun getOpponent(): Trainer {
-    return if (currentTrainer == trainerOne) trainerTwo else trainerOne
-  }
+        private fun getOpponent(): Trainer {
+                return if (currentTrainer == trainerOne) trainerTwo else trainerOne
+        }
 
-  fun takeTurn(attack: Attack): Monster {
-    if (status == BattleStatus.FINISHED) {
-      throw IllegalStateException("Battle is already finished")
-    }
-    if (currentTrainer == Trainer.EMPTY) {
-      throw IllegalStateException(
-        "Trainer needs to have been set for this operation"
-      )
-    }
-    val monster = currentTrainer.nextBattleReadyMonster()
-    val opponent = getOpponent()
-    val opponentMonster = opponent.nextBattleReadyMonster()
-    monster.attack(attack, opponentMonster)
-    currentTrainer = opponent
+        override fun takeTurn(attack: Attack): Monster {
+                if (status == BattleStatus.FINISHED) {
+                        throw IllegalStateException("Battle is already finished")
+                }
+                if (currentTrainer == Trainer.EMPTY) {
+                        throw IllegalStateException(
+                                "Trainer needs to have been set for this operation"
+                        )
+                }
+                val monster = currentTrainer.nextBattleReadyMonster()
+                val opponent = getOpponent()
+                val opponentMonster = opponent.nextBattleReadyMonster()
+                monster.attack(attack, opponentMonster)
+                currentTrainer = opponent
 
-    advanceRound()
-    return opponentMonster
-  }
+                advanceAndSaveRound()
+                return opponentMonster
+        }
 
-  fun determineWinner(): Trainer {
-    requireNoTrainerIsEmpty()
-    if (trainerOne.isDefeated()) return trainerTwo
-    if (trainerTwo.isDefeated()) return trainerOne
-    return Trainer.EMPTY
-  }
+        override fun determineWinner(): Trainer {
+                requireNoTrainerIsEmpty()
+                if (trainerOne.isDefeated()) return trainerTwo
+                if (trainerTwo.isDefeated()) return trainerOne
+                return Trainer.EMPTY
+        }
 
-  private fun requireNoTrainerIsEmpty() {
-    if (trainerOne == Trainer.EMPTY || trainerTwo == Trainer.EMPTY) {
-      throw EmptyTrainerException()
-    }
-  }
+        private fun requireNoTrainerIsEmpty() {
+                if (trainerOne == Trainer.EMPTY || trainerTwo == Trainer.EMPTY) {
+                        throw EmptyTrainerException()
+                }
+        }
 
-  class EmptyTrainerException : Exception("Trainer needs to have been set for this operation") {
-  }
+        class EmptyTrainerException :
+                Exception("Trainer needs to have been set for this operation") {}
 
-  enum class BattleStatus {
-    PREGAME,
-    STARTED,
-    FINISHED,
-  }
-
-  override fun toString(): String {
-    return if (trainerOne != Trainer.EMPTY &&
-      trainerTwo != Trainer.EMPTY &&
-      currentTrainer != Trainer.EMPTY
-    )
-      """Battle ($battleId):
+        override fun toString(): String {
+                return if (trainerOne != Trainer.EMPTY &&
+                                trainerTwo != Trainer.EMPTY &&
+                                currentTrainer != Trainer.EMPTY
+                )
+                        """Battle ($battleId):
 ${trainerOne.name} vs. ${trainerTwo.name}
 Round: $currentRound
 Next Attacker: ${currentTrainer.name}"""
-    else "Battle with ID: $battleId is in a pregame state"
-  }
+                else "Battle with ID: $battleId is in a pregame state"
+        }
 }
