+++ src/main/kotlin/hwr/oop/tnp/core/BattleStatus.kt
@@ -1,0 +1,7 @@
+package hwr.oop.tnp.core
+
+enum class BattleStatus {
+        PREGAME,
+        STARTED,
+        FINISHED,
+}
