+++ src/main/kotlin/hwr/oop/tnp/persistency/SaveBattlePort.kt
@@ -2,6 +2,6 @@ package hwr.oop.tnp.persistency
 
 import hwr.oop.tnp.core.Battle
 
-interface GamePersistencePort {
-  fun saveBattle(battle: Battle)
+interface SaveBattlePort {
+        fun saveBattle(battle: Battle)
 }
